GameBoard class created, implementation started;
Created GameBoard class to hold an array of ships on the board and a pseudo dictionary to hold missed hit coordinates
Implemented initShip() function which checks against incorrect lengths vs # of provided coordinates, as well as out of bound coordinates. Creates new ship and adds it to shipArr.
Implmented receiveAttack() function which receives an x & y coordinate, checks shipArr to see if the coordinates hit any ships. if yes, the ships hit() function is called. if no, the coordinates are added to the missArr dictionary.